Example encoder for embedded and temporal layers
Use the existing scalable_encoder,
remove the old frame flags, add temporal
and embedded layers.
