+++ examples/scalable_encoder.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, Alliance for Open Media. All rights reserved
+ * Copyright (c) 2026, Alliance for Open Media. All rights reserved
  *
  * This source code is subject to the terms of the BSD 3-Clause Clear License
  * and the Alliance for Open Media Patent License 1.0. If the BSD 3-Clause Clear
@@ -9,66 +9,6 @@
  * source code in the PATENTS file, you can obtain it at
  * aomedia.org/license/patent-license/.
  */
-
-// Scalable Encoder
-// ==============
-//
-// This is an example of a scalable encoder loop. It takes two input files in
-// YV12 format, passes it through the encoder, and writes the compressed
-// frames to disk in OBU format.
-//
-// Getting The Default Configuration
-// ---------------------------------
-// Encoders have the notion of "usage profiles." For example, an encoder
-// may want to publish default configurations for both a video
-// conferencing application and a best quality offline encoder. These
-// obviously have very different default settings. Consult the
-// documentation for your codec to see if it provides any default
-// configurations. All codecs provide a default configuration, number 0,
-// which is valid for material in the vacinity of QCIF/QVGA.
-//
-// Updating The Configuration
-// ---------------------------------
-// Almost all applications will want to update the default configuration
-// with settings specific to their usage. Here we set the width and height
-// of the video file to that specified on the command line. We also scale
-// the default bitrate based on the ratio between the default resolution
-// and the resolution specified on the command line.
-//
-// Encoding A Frame
-// ----------------
-// The frame is read as a continuous block (size = width * height * 3 / 2)
-// from the input file. If a frame was read (the input file has not hit
-// EOF) then the frame is passed to the encoder. Otherwise, a NULL
-// is passed, indicating the End-Of-Stream condition to the encoder. The
-// `frame_cnt` is reused as the presentation time stamp (PTS) and each
-// frame is shown for one frame-time in duration. The flags parameter is
-// unused in this example.
-
-// Forced Keyframes
-// ----------------
-// Keyframes can be forced by setting the AVM_EFLAG_FORCE_KF bit of the
-// flags passed to `avm_codec_control()`. In this example, we force a
-// keyframe every <keyframe-interval> frames. Note, the output stream can
-// contain additional keyframes beyond those that have been forced using the
-// AVM_EFLAG_FORCE_KF flag because of automatic keyframe placement by the
-// encoder.
-//
-// Processing The Encoded Data
-// ---------------------------
-// Each packet of type `AVM_CODEC_CX_FRAME_PKT` contains the encoded data
-// for this frame. We write a IVF frame header, followed by the raw data.
-//
-// Cleanup
-// -------
-// The `avm_codec_destroy` call frees any memory allocated by the codec.
-//
-// Error Handling
-// --------------
-// This example does not special case any error return codes. If there was
-// an error, a descriptive message is printed and the program exits. With
-// few exeptions, avm_codec functions return an enumerated error status,
-// with the value `0` indicating success.
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -84,9 +24,10 @@ static const char *exec_name;
 
 void usage_exit(void) {
   fprintf(stderr,
-          "Usage: %s <codec> <width> <height> <infile0> <infile1> "
+          "Usage: %s <codec> <width> <height> <infile0>  "
           "<outfile> <frames to encode>\n"
-          "See comments in scalable_encoder.c for more information.\n",
+          "See comments in embedded_temporal_layers_encoder.c for more "
+          "information.\n",
           exec_name);
   exit(EXIT_FAILURE);
 }
@@ -121,22 +62,19 @@ static int encode_frame(avm_codec_ctx_t *codec, avm_image_t *img,
 
 int main(int argc, char **argv) {
   FILE *infile0 = NULL;
-  FILE *infile1 = NULL;
   avm_codec_enc_cfg_t cfg;
   int frame_count = 0;
-  avm_image_t raw0, raw1;
+  avm_image_t raw0;
   avm_codec_err_t res;
   AvxVideoInfo info;
-  const int fps = 30;
-  const int bitrate = 200;
   int keyframe_interval = 0;
   int max_frames = 0;
   int frames_encoded = 0;
+  const int fps = 30;
   const char *codec_arg = NULL;
   const char *width_arg = NULL;
   const char *height_arg = NULL;
   const char *infile0_arg = NULL;
-  const char *infile1_arg = NULL;
   const char *outfile_arg = NULL;
   //  const char *keyframe_interval_arg = NULL;
   FILE *outfile = NULL;
@@ -147,15 +85,14 @@ int main(int argc, char **argv) {
   // "missing-field-initializers" warning in some compilers.
   memset(&info, 0, sizeof(info));
 
-  if (argc != 8) die("Invalid number of arguments");
+  if (argc != 7) die("Invalid number of arguments");
 
   codec_arg = argv[1];
   width_arg = argv[2];
   height_arg = argv[3];
   infile0_arg = argv[4];
-  infile1_arg = argv[5];
-  outfile_arg = argv[6];
-  max_frames = (int)strtol(argv[7], NULL, 0);
+  outfile_arg = argv[5];
+  max_frames = (int)strtol(argv[6], NULL, 0);
 
   avm_codec_iface_t *encoder = get_avm_encoder_by_short_name(codec_arg);
   if (!encoder) die("Unsupported codec.");
@@ -173,18 +110,12 @@ int main(int argc, char **argv) {
 
   if (!avm_img_alloc(&raw0, AVM_IMG_FMT_I420, info.frame_width,
                      info.frame_height, 1)) {
-    die("Failed to allocate image for layer 0.");
-  }
-  if (!avm_img_alloc(&raw1, AVM_IMG_FMT_I420, info.frame_width,
-                     info.frame_height, 1)) {
-    die("Failed to allocate image for layer 1.");
+    die("Failed to allocate image.");
   }
 
   //  keyframe_interval = (int)strtol(keyframe_interval_arg, NULL, 0);
-  keyframe_interval = 100;
+  keyframe_interval = 1000;
   if (keyframe_interval < 0) die("Invalid keyframe interval value.");
-
-  printf("Using %s\n", avm_codec_iface_name(encoder));
 
   avm_codec_ctx_t codec;
   res = avm_codec_enc_config_default(encoder, &cfg, 0);
@@ -194,77 +125,117 @@ int main(int argc, char **argv) {
   cfg.g_h = info.frame_height;
   cfg.g_timebase.num = info.time_base.numerator;
   cfg.g_timebase.den = info.time_base.denominator;
-  cfg.rc_target_bitrate = bitrate;
+  cfg.rc_end_usage = AVM_Q;
+  cfg.rc_min_quantizer = 150;
+  cfg.rc_max_quantizer = 150;
   cfg.g_error_resilient = 0;
   cfg.g_lag_in_frames = 0;
-  cfg.rc_end_usage = AVM_Q;
   cfg.signal_td = 0;
   outfile = fopen(outfile_arg, "wb");
   if (!outfile) die("Failed to open %s for writing.", outfile_arg);
 
   if (!(infile0 = fopen(infile0_arg, "rb")))
     die("Failed to open %s for reading.", infile0_arg);
-  if (!(infile1 = fopen(infile1_arg, "rb")))
-    die("Failed to open %s for reading.", infile0_arg);
 
   if (avm_codec_enc_init(&codec, encoder, &cfg, 0))
     die("Failed to initialize encoder");
-  if (avm_codec_control(&codec, AVME_SET_CPUUSED, 8))
-    die_codec(&codec, "Failed to set cpu to 8");
+  if (avm_codec_control(&codec, AVME_SET_CPUUSED, 5))
+    die_codec(&codec, "Failed to set cpu to 5");
 
-  if (avm_codec_control(&codec, AV2E_SET_TILE_COLUMNS, 2))
-    die_codec(&codec, "Failed to set tile columns to 2");
-  if (avm_codec_control(&codec, AV2E_SET_NUM_TG, 3))
-    die_codec(&codec, "Failed to set num of tile groups to 3");
+  // Test case: currently only (1, 2), (2, 1), (2, 2), (1, 3), (3 1),
+  // more cases will be added.
+  int num_embedded_layers = 2;
+  int num_temporal_layers = 1;
 
-  if (avm_codec_control(&codec, AVME_SET_NUMBER_MLAYERS, 2))
-    die_codec(&codec, "Failed to set number of spatial layers to 2");
+  if (avm_codec_control(&codec, AVME_SET_NUMBER_MLAYERS, num_embedded_layers))
+    die_codec(&codec, "Failed to set number of embedded layers.");
+  if (avm_codec_control(&codec, AVME_SET_NUMBER_TLAYERS, num_temporal_layers))
+    die_codec(&codec, "Failed to set number of temporal layers.");
 
   // Encode frames.
   while (avm_img_read(&raw0, infile0)) {
     int flags = 0;
 
-    // configure and encode base layer
-
-    if (keyframe_interval > 0 && frames_encoded % keyframe_interval == 0)
+    if (keyframe_interval > 0 &&
+        frames_encoded % (keyframe_interval * num_embedded_layers) == 0) {
       flags |= AVM_EFLAG_FORCE_KF;
-    else
-      // use previous base layer (LAST) as sole reference
-      // save this frame as LAST to be used as reference by enhanmcent layer
-      // and next base layer
-      flags |= AVM_EFLAG_NO_REF_LAST2 | AVM_EFLAG_NO_REF_LAST3 |
-               AVM_EFLAG_NO_REF_GF | AVM_EFLAG_NO_REF_ARF |
-               AVM_EFLAG_NO_REF_BWD | AVM_EFLAG_NO_REF_ARF2;
-    cfg.g_w = info.frame_width;
-    cfg.g_h = info.frame_height;
-    if (avm_codec_enc_config_set(&codec, &cfg))
-      die_codec(&codec, "Failed to set enc cfg for layer 0");
-    if (avm_codec_control(&codec, AVME_SET_MLAYER_ID, 0))
-      die_codec(&codec, "Failed to set layer id to 0");
-    if (avm_codec_control(&codec, AVME_SET_QP, 62))
-      die_codec(&codec, "Failed to set cq level");
-    encode_frame(&codec, &raw0, frame_count++, flags, outfile);
+    }
 
-    // configure and encode enhancement layer
+    // For embedded layers: call the encoder num_embedded_layers times with same
+    // input at different scales. So the example here is spatial layers.
+    for (int sl = 0; sl < num_embedded_layers; sl++) {
+      // Add more cases and move/refactor, up to (3,3).
+      if (num_temporal_layers == 2 && num_embedded_layers == 1) {
+        if (frames_encoded % 2 == 0) {
+          avm_codec_control(&codec, AVME_SET_TLAYER_ID, 0);
+        } else {
+          avm_codec_control(&codec, AVME_SET_TLAYER_ID, 1);
+        }
+      } else if (num_temporal_layers == 1 && num_embedded_layers == 2) {
+        if (frames_encoded % 2 == 0) {
+          struct avm_scaling_mode mode = { AVME_ONETWO, AVME_ONETWO };
+          avm_codec_control(&codec, AVME_SET_SCALEMODE, &mode);
+          avm_codec_control(&codec, AVME_SET_MLAYER_ID, 0);
+        } else {
+          struct avm_scaling_mode mode = { AVME_NORMAL, AVME_NORMAL };
+          avm_codec_control(&codec, AVME_SET_SCALEMODE, &mode);
+          avm_codec_control(&codec, AVME_SET_MLAYER_ID, 1);
+        }
+      } else if (num_temporal_layers == 2 && num_embedded_layers == 2) {
+        if (frames_encoded % 4 == 0) {
+          struct avm_scaling_mode mode = { AVME_ONETWO, AVME_ONETWO };
+          avm_codec_control(&codec, AVME_SET_SCALEMODE, &mode);
+          avm_codec_control(&codec, AVME_SET_MLAYER_ID, 0);
+          avm_codec_control(&codec, AVME_SET_TLAYER_ID, 0);
+        } else if (frames_encoded % 2 == 0) {
+          struct avm_scaling_mode mode = { AVME_ONETWO, AVME_ONETWO };
+          avm_codec_control(&codec, AVME_SET_SCALEMODE, &mode);
+          avm_codec_control(&codec, AVME_SET_MLAYER_ID, 0);
+          avm_codec_control(&codec, AVME_SET_TLAYER_ID, 1);
+        } else if ((frames_encoded - 1) % 4 == 0) {
+          avm_codec_control(&codec, AVME_SET_MLAYER_ID, 1);
+          avm_codec_control(&codec, AVME_SET_TLAYER_ID, 0);
+        } else if ((frames_encoded - 1) % 2 == 0) {
+          avm_codec_control(&codec, AVME_SET_MLAYER_ID, 1);
+          avm_codec_control(&codec, AVME_SET_TLAYER_ID, 1);
+        }
+      } else if (num_temporal_layers == 3 && num_embedded_layers == 1) {
+        if (frames_encoded % 4 == 0) {
+          avm_codec_control(&codec, AVME_SET_MLAYER_ID, 0);
+          avm_codec_control(&codec, AVME_SET_TLAYER_ID, 0);
+        } else if (frames_encoded % 2 == 0) {
+          avm_codec_control(&codec, AVME_SET_MLAYER_ID, 0);
+          avm_codec_control(&codec, AVME_SET_TLAYER_ID, 1);
+        } else {
+          avm_codec_control(&codec, AVME_SET_MLAYER_ID, 0);
+          avm_codec_control(&codec, AVME_SET_TLAYER_ID, 2);
+        }
+      } else if (num_temporal_layers == 1 && num_embedded_layers == 3) {
+        if (frames_encoded % 3 == 0) {
+          struct avm_scaling_mode mode = { AVME_ONEFOUR, AVME_ONEFOUR };
+          avm_codec_control(&codec, AVME_SET_SCALEMODE, &mode);
+          avm_codec_control(&codec, AVME_SET_MLAYER_ID, 0);
+          avm_codec_control(&codec, AVME_SET_TLAYER_ID, 0);
+        } else if ((frames_encoded - 1) % 3 == 0) {
+          struct avm_scaling_mode mode = { AVME_ONETWO, AVME_ONETWO };
+          avm_codec_control(&codec, AVME_SET_SCALEMODE, &mode);
+          avm_codec_control(&codec, AVME_SET_MLAYER_ID, 1);
+          avm_codec_control(&codec, AVME_SET_TLAYER_ID, 0);
+        } else if ((frames_encoded - 2) % 3 == 0) {
+          struct avm_scaling_mode mode = { AVME_NORMAL, AVME_NORMAL };
+          avm_codec_control(&codec, AVME_SET_SCALEMODE, &mode);
+          avm_codec_control(&codec, AVME_SET_MLAYER_ID, 2);
+          avm_codec_control(&codec, AVME_SET_TLAYER_ID, 0);
+        }
+      }
 
-    //  use LAST (base layer) as sole reference
-    flags = AVM_EFLAG_NO_REF_LAST2 | AVM_EFLAG_NO_REF_LAST3 |
-            AVM_EFLAG_NO_REF_GF | AVM_EFLAG_NO_REF_ARF | AVM_EFLAG_NO_REF_BWD |
-            AVM_EFLAG_NO_REF_ARF2 | AVM_EFLAG_NO_UPD_ALL;
-    cfg.g_w = info.frame_width;
-    cfg.g_h = info.frame_height;
-    avm_img_read(&raw1, infile1);
-    if (avm_codec_enc_config_set(&codec, &cfg))
-      die_codec(&codec, "Failed to set enc cfg for layer 1");
-    if (avm_codec_control(&codec, AVME_SET_MLAYER_ID, 1))
-      die_codec(&codec, "Failed to set layer id to 1");
-    if (avm_codec_control(&codec, AVME_SET_QP, 10))
-      die_codec(&codec, "Failed to set cq level");
-    encode_frame(&codec, &raw1, frame_count++, flags, outfile);
+      encode_frame(&codec, &raw0, frame_count++, flags, outfile);
 
-    frames_encoded++;
+      frames_encoded++;
+    }
 
-    if (max_frames > 0 && frames_encoded >= max_frames) break;
+    if (max_frames > 0 && frames_encoded >= max_frames * num_embedded_layers)
+      break;
   }
 
   // Flush encoder.
@@ -272,11 +243,9 @@ int main(int argc, char **argv) {
 
   printf("\n");
   fclose(infile0);
-  fclose(infile1);
-  printf("Processed %d frames.\n", frame_count / 2);
+  printf("Processed %d frames.\n", frames_encoded);
 
   avm_img_free(&raw0);
-  avm_img_free(&raw1);
   if (avm_codec_destroy(&codec)) die_codec(&codec, "Failed to destroy codec.");
 
   fclose(outfile);
